START SEM

+++ dailyoptapi/src/main/java/routingdelivery/smartlog/brenntag/service/ItemAmount.java
@@ -2,20 +2,20 @@ package routingdelivery.smartlog.brenntag.service;
 
 public class ItemAmount {
 	public int itemIndex;
-	public int amount;
+	public double amount;
 	public int getItemIndex() {
 		return itemIndex;
 	}
 	public void setItemIndex(int itemIndex) {
 		this.itemIndex = itemIndex;
 	}
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
-	public void setAmount(int amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
-	public ItemAmount(int itemIndex, int amount) {
+	public ItemAmount(int itemIndex, double amount) {
 		super();
 		this.itemIndex = itemIndex;
 		this.amount = amount;
